# apps/banking-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Intergalactic Banking API
  version: 2.0.0
  description: A deterministic, run-scoped Banking API for comparing direct agent integrations with Postman
    Fabric Gateway. Monetary values are non-negative safe integers in minor units; all fictional currencies
    currently use exponent 0.
  contact:
    name: Intergalactic Banking API team
    url: https://www.postman.com/
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local demo server
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/customers/me:
    get:
      tags:
        - Customers
      summary: Get the current customer
      description: Get the current customer in the selected run.
      operationId: getCurrentCustomer
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Get the current customer succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    patch:
      tags:
        - Customers
      summary: Update customer contact details
      description: Update customer contact details in the selected run.
      operationId: updateCurrentCustomer
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Update customer contact details succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
  /api/v1/beneficiaries:
    get:
      tags:
        - Beneficiaries
      summary: List beneficiaries
      description: List beneficiaries in the selected run.
      operationId: listBeneficiaries
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List beneficiaries succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeneficiaryList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags:
        - Beneficiaries
      summary: Create an unverified beneficiary
      description: Create an unverified beneficiary in the selected run.
      operationId: createBeneficiary
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Create an unverified beneficiary succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeneficiaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeneficiaryRequest"
  /api/v1/beneficiaries/{id}:
    get:
      tags:
        - Beneficiaries
      summary: Get a beneficiary
      description: Get a beneficiary in the selected run.
      operationId: getBeneficiary
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Get a beneficiary succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeneficiaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    patch:
      tags:
        - Beneficiaries
      summary: Update a beneficiary
      description: Update a beneficiary in the selected run.
      operationId: updateBeneficiary
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Update a beneficiary succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeneficiaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBeneficiaryRequest"
    delete:
      tags:
        - Beneficiaries
      summary: Deactivate a beneficiary
      description: Deactivate a beneficiary in the selected run.
      operationId: cancelBeneficiary
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Deactivate a beneficiary succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeneficiaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/cards:
    get:
      tags:
        - Cards
      summary: List cards
      description: List cards in the selected run.
      operationId: listCards
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List cards succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags:
        - Cards
      summary: Create a card
      description: Create a card in the selected run.
      operationId: createCard
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Create a card succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
  /api/v1/cards/{id}:
    get:
      tags:
        - Cards
      summary: Get a card
      description: Get a card in the selected run.
      operationId: getCard
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Get a card succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/cards/{id}/freeze:
    post:
      tags:
        - Cards
      summary: Freeze an active card
      description: Freeze an active card in the selected run.
      operationId: freezeCard
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Freeze an active card succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/cards/{id}/unfreeze:
    post:
      tags:
        - Cards
      summary: Unfreeze a frozen card
      description: Unfreeze a frozen card in the selected run.
      operationId: unfreezeCard
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Unfreeze a frozen card succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/cards/{id}/limits:
    patch:
      tags:
        - Cards
      summary: Update a card spending limit
      description: Update a card spending limit in the selected run.
      operationId: updateCardLimit
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Update a card spending limit succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardLimitRequest"
  /api/v1/cards/{id}/replace:
    post:
      tags:
        - Cards
      summary: Replace an active or frozen card
      description: Replace an active or frozen card in the selected run.
      operationId: replaceCard
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: id
          in: path
          required: true
          description: Run-scoped id value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Replace an active or frozen card succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/scheduled-payments:
    get:
      tags:
        - Payments
      summary: List scheduled payments
      description: List scheduled payments in the selected run.
      operationId: listScheduledPayments
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List scheduled payments succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledPaymentList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags:
        - Payments
      summary: Create a scheduled payment
      description: Create a scheduled payment in the selected run.
      operationId: createScheduledPayment
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Create a scheduled payment succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduledPaymentRequest"
...
  /api/v1/standing-orders:
    get:
      tags:
        - Payments
      summary: List standing orders
      description: List standing orders in the selected run.
      operationId: listStandingOrders
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List standing orders succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandingOrderList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags:
        - Payments
      summary: Create a standing order
      description: Create a standing order in the selected run.
      operationId: createStandingOrder
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Create a standing order succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStandingOrderRequest"
...
  /api/v1/statements:
    get:
      tags:
        - Statements
      summary: List statement transactions
      description: List statement transactions in the selected run.
      operationId: listStatements
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List statement transactions succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/audit-events:
    get:
      tags:
        - Audit
      summary: List append-only audit events
      description: List append-only audit events in the selected run.
      operationId: listAuditEvents
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List append-only audit events succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: A credential resolving to a CUSTOMER or ADMIN principal.
  parameters:
    DemoRunId:
      name: X-Demo-Run-Id
      in: header
      required: false
      description: Isolates state for a comparison lane; defaults to default.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]{1,64}$
        default: default
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Caller-provided correlation identifier; generated deterministically when omitted.
      schema:
        type: string
        minLength: 1
...
  schemas:
...
    CustomerResponse:
      type: object
      additionalProperties: false
      required:
        - customer
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
...
  responses:
    BadRequest:
      description: The request is malformed or violates a schema.
      headers: &a10
        X-Demo-Run-Id: *a1
        X-Request-Id: *a2
        X-RateLimit-Limit: *a3
        X-RateLimit-Remaining: *a4
        X-RateLimit-Reset: *a5
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: A valid API key is required.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...